+++ runtime/tests/Makefile
@@ -1,5 +1,5 @@
 # Make sure these are in .gitignore
-TESTPROGS=string_hash string_cmp string_concat string_len immediate_string tagged_ptr float_print error_print error_backtrace1 error_backtrace2 demangle list_construct decimal_operations
+TESTPROGS=string_hash string_cmp string_concat string_len immediate_string tagged_ptr float_print error_print error_backtrace1 error_backtrace2 demangle list_construct decimal_arith
 # following will test balrt_inline, and must be named inlinert_*
 TESTINLINERTPROGS=inlinert_float inlinert_int inlinert_string inlinert_list
 TESTPROG_STAMPS=$(addsuffix .stamp, $(TESTPROGS) $(TESTINLINERTPROGS))
